Modified today label, removed Javadoc errors.
The today label was not closing the pop-up (like the clear label does),
and it also was not selecting the current date if the date picker was
null. Modified the today label action to fix these things.

Also made minor modifications to a few javadoc entries that were causing
compile errors (failure to compile) under Netbeans and Maven.