Bug fix: Month selector not selecting the correct month in certain cases.
Previous behavior: if you select the last day of the month with a long
number of days (for example March 31), and then use the month drop down
to select a month with a shorter number of days (for example February),
the date selector would switch to a day that is not inside of February.

Fixed this behavior in AbstractDateModel so that the month drop-down
will work correctly in these cases.